+++ src/pages/Main/MainPage.tsx
@@ -1,5 +1,3 @@
-import BottomNav from '@/components/BottomNav';
-
 function MainPage() {
    return <p>메인페이지</p>;
 }
