+++ src/pages/Recommend/TeaListPage.tsx
@@ -1,5 +1,3 @@
-import BottomNav from '@/components/BottomNav';
-
 function TeaListPage() {
    return <p>추천 티 리스트 페이지</p>;
 }
