+++ src/pages/Settings/MyPage.tsx
@@ -1,5 +1,3 @@
-import BottomNav from '@/components/BottomNav';
-
 function MyPage() {
    return <p>마이페이지</p>;
 }
